requirements.txt
pip==23.2.1  
setuptools==68.0.0  
wheel==0.41.0  
alembic>=1.10.0  
aniso8601>=10.0.0  
APIFlask>=2.3.2  
apispec>=6.3.0  
APScheduler>=3.11.0  
bcrypt>=4.0.1  
bidict>=0.23.1  
blinker>=1.9.0  
Brotli>=1.1.0  
browser-history>=0.2.0  
click>=8.1.3  
colorama>=0.4.6  
configparser>=6.0.0  
coverage>=5.3.1,<6.0.0  
deprecation>=2.1.0  
dnspython>=2.3.0  
eyed3>=0.9.7  
filetype>=1.2.0  
Flask>=2.3.0,<3.0.0  
Flask-Admin>=1.6.1  
Flask-APScheduler>=1.13.1  
Flask-Bcrypt>=1.0.1  
Flask-Caching>=1.10.1  
Flask-Compress>=1.17  
Flask-Cors>=5.0.1  
Flask-DebugToolbar>=0.16.0  
Flask-HTTPAuth>=4.8.0  
Flask-JSON>=0.4.0  
Flask-JWT-Extended>=4.7.1  
Flask-Limiter>=3.10.1  
Flask-Login>=0.6.3  
Flask-Mail>=0.10.0  
flask-marshmallow>=1.3.0  
Flask-Migrate>=4.1.0  
Flask-RESTful>=0.3.10  
Flask-SocketIO>=5.5.1   
Flask-SQLAlchemy>=3.0.0  
Flask-Testing>=0.8.1  
Flask-WTF>=1.0.1,<2.0.0  
greenlet>=2.0.2  
gunicorn>=23.0.0  
h11>=0.14.0  
idna>=3.4  
iniconfig>=2.0.0  
itsdangerous>=2.1.2  
Jinja2>=3.0.0,<4.0.0  
limits>=3.5.0  
Mako>=1.1.6  
markdown-it-py>=3.0.0  
MarkupSafe>=2.1.3  
marshmallow>=3.13.0,<4.0.0  
mdurl>=0.1.2  
ordered-set>=4.1.0  
packaging>=21.0  
pip>=23.0  
pluggy>=1.0.0  
pycryptodome>=3.15.0  
Pygments>=2.15.1  
PyJWT>=2.0.0,<3.0.0  
pytest>=7.0.0,<8.0.0  
python-dateutil>=2.8.2,<3.0.0  
python-dotenv==1.0.0  
python-engineio>=4.11.0  
python-socketio>=5.12.0  
pytz>=2021.3  
rich>=12.0.0  
setuptools>=57.5.0  
simple-websocket==1.1.0  
six>=1.15.0  
SQLAlchemy>=1.4.31,<2.0.0  
toml>=0.10.2  
typing_extensions>=3.10.0  
tzdata>=2021.5  
tzlocal>=5.2  
webargs==8.6.0  
Werkzeug>=2.2.0,<3.0.0  
wheel>=0.36.0  
wrapt>=1.12.1  
wsproto==1.2.0  
WTForms>=3.0.0,<4.0.0  
zstandard>=0.16.0 